pyproject.toml
[project]
name = "geo-mcp"
version = "0.1.0"
description = "UK-specialist geospatial MCP server"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "asyncpg>=0.29",
    "pyproj>=3.6",
    "python-dotenv>=1.0",
    "rasterio>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
geo-mcp = "geo_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["geo_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "PL"]
ignore = ["PLR0913"]
